Prototype listening on SocketAddr instead of port
The basic idea is the same, but it better mimics
the normal socket interface.

The Switchboard can now be thought of as
a parody of the entire internet
instead of a single machine.

Most everything is a direct mapping,
but there are two uncertain situations:
 *  When connect() is called, there is no clear indicator
    as to where the client is connecting *from*.
    So if we ever wanted to implement e.g. peer_addr
    on a MemorySocket,
    the behaviour would be undefined.
 *  It is not clear what address to listen on when
     connecting to 0.0.0.0.

Both of these *could* be solved by mapping to the local
machine address (using e.g the get_if_addrs crate),
but I'm not sure that make sense.
Another option I toyed with was having a "set_connect_address"
global, but that simply lacked elegance.

( Discussion in #3 )